Update test and add the capacity validation when drone is loading medication

+++ src/main/java/com/hitachi/test/constant/Constant.java
@@ -5,6 +5,6 @@ public interface Constant {
   String DATA = "data";
   String DRONE = "drone";
   int DEFAULT_DELIVERY_BATTERY_REQUIRED = 25;
-  double defaultWeight = 1000;
+  int defaultWeight = 1000;
   int DEFAULT_DELIVERY_BATTERY_COST = 10;
 }
